build: new backend with typescript
I have greatly enjoyed working in perl, but I am having too much trouble with
mojolicious debugging and configuration

I had the same trouble with dancer2, and thus have tried both of major
frameworks for serving content in perl.  Both are optimized for websites
and not backend services *or* for home grown systems and do not provide
a robust skeleton that I can populate.