feat: generals tables
on adding in the attacking values, the table slowed to a crawl on each update.
it will get worse with reinforcing, and wall use cases, even if I totally
special case mayors.
I need to dynamically generate the columns based on the use case.
this sets up for that, though I have not started to manipulate the table
itself yet.  I have only set up the value selector so far.