feat: genearls table:
performance boost from populating the nano store in astro, the component wasn't reusable anyway
trying to *make* the component reusable, and failing.
the columns are static once set for the table with few exceptions.
this forces me to rethink a number of things, only some of
which are working now.