feat: forms
using @lit-labs state as a file full of primatives.
I remember considering doing this with atoms, but this uses lit-native types.

This seems to avoid the import map mess.

it should work well when I can convert to tanstack with the lit adapter.
I'm unsure how it will work with the vanilla js tanstack I have now.