feat: reactive monster reports replacement
after way too much work, the first cell of the results table
is reactive to changes in the other two tables.

This required radically rethinking how I store state for the tables.

My initial thought to allow TanStack to store its own state was
massively inadequate, and it took several tries to figure out how
to work with the current state of @lit-labs/signals, signals-utils,
and the signals proposal they are based on.

I still need to flesh out the results table, both completing the
computation for attack and filling in the rest of the cells.