feat: reactive monster reports replacement
the modifier is insufficiently reactive,
some of the logic needs to move from reference.ts to state.ts
and the call to trigger it needs to be because the
SimulatorState.modifier attribute is itself a computed signal,
not a standard signal.

I also hate that I am using home grown csv parsing because I tried
3 different csv parsing libraries and none worked right with greenwood
(with different errors).  Some of these might have worked if I had been
willing to use the fetch approach that I am currently using,
but I did not go back and retry them after I had to give up
on using the greenwood raw import plugin because it strips newlines
out of the files. Without newlines, I cannot find the row ends.

I also would rather some slightly more robust fetch code, but
this should be sufficient.