fix: allow the standard layout in serve mode
this was a duplicative import, I also have a link tag importing it, and
that was the operative import statement.  At serve time, the code gets
"sucked in" (for lack of a better phrase/wording) to the file that
includes this layout and the relative path to the sidebar is different
than what it is from the code on my filesystem.  Thus the import works
in build and dev mode, but not in serve mode.  When the import fails,
the page crashes.